Add checkbox to use image "alt" attribute for file name
Some image elements have meaningless file names but descriptive "alt"
image attributes, for example:

  <img src="http://example.com/620918775623.jpg" alt="Image17"/>

This commit adds the option to use the image's "alt" attribute for the
file name, and optionally add a filename extension (like ".jpg"), so
that "620918775623.jpg" may be saved as "Image17.jpg".

If the option is enabled but there is no "alt" attribute, the original
behavior is preserved.